refactor: Вынес протоколы репозиториев в дирректорию store в модуле matrix
