+++ app/infra/ioc/providers/matrix/repo.py
@@ -1,8 +1,8 @@
 from dishka import provide
 from dishka import Provider
 from dishka import Scope
-from matrix.action.store.user_matrix import IUserMatrixRepository
-from matrix.model.repo.user_matrix import UserMatrixRepository
+from matrix.store.proto.user_matrix import IUserMatrixRepository
+from matrix.store.repo.user_matrix import UserMatrixRepository
 
 
 class UserMatrixRepoProvider(Provider):
