+++ app/matrix/store/repo/user_matrix.py
@@ -1,9 +1,9 @@
 from typing import List
 
-from matrix.action.store.user_matrix import IUserMatrixRepository
 from matrix.domain.matrix import Matrix as MatrixDomain
-from matrix.model.entity.matrix import Matrix
-from matrix.model.entity.user_matrix import UserMatrix
+from matrix.store.entity.matrix import Matrix
+from matrix.store.entity.user_matrix import UserMatrix
+from matrix.store.proto.user_matrix import IUserMatrixRepository
 from sqlalchemy.ext.asyncio.session import AsyncSession
 from sqlalchemy.sql import select
 from user.domain.user import User
